metodos create y delete

+++ am-core-web-service/src/main/java/org/am/core/web/dto/admingeneral/LevelRequest.java
@@ -2,7 +2,7 @@ package org.am.core.web.dto.admingeneral;
 
 import java.util.List;
 
-public record LevelRequest(Integer levelIdentifier,
+public record LevelRequest(Short levelIdentifier,
                            String levelName,
                            List<SubjectCurriculumRequest> subjectCurriculumList) {
 }
